Refactor DefaultAnnotationPageLabel
